pyproject.toml
[project]
name = "artisendapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Brit Beaty",email = "britlbeaty@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
